Ratchet: Don't permanently store private_identity + derive chain key
This makes ratchet_create derive a initial root_key *and* a chain key
which is the same for send/receive. Only one of them will be used though
(depending on who starts the conversation, if we start it, only the send
key is used, only the receive key otherwise).

This deviates from the axolotl protocol so this has to be evaluated or
changed later.